requirements.txt
#prod
Django==3.0.8
gunicorn==20.0.4
whitenoise==5.1.0
dj-database-url==0.5.0
django-graphql-jwt==0.3.0
django_graphql_auth==0.3.11
graphene-django==2.12.1
django-phonenumber-field==4.0.0
phonenumberslite==8.12.7
django-eventtools==1.0.0

# dev and test
black==19.3b0
flake8==3.7.7
isort==4.3.20